“他ユーザーの遷移リンクまで”

+++ resources/views/posts/index.blade.php
@@ -1,37 +1,52 @@
 @extends('layouts.login')
 
 @section('content')
-<h2>機能を実装していきましょう。</h2>
+
 {!! Form::open(['url' => 'post/create']) !!}
 <!-- HTMLではフォームタグは長くなるがフォームファサードで短く書き換えられている -->
-     <div class="form-group">
-         {!! Form::input('text', 'newPost', null, ['required', 'class' => 'form-control', 'placeholder' => '投稿内容を入力してください']) !!}
-     </div>
-     <button type="submit" class="btn btn-success pull-right"><img class="Upload" src="images/post.png"></button>
+    <div class= "comment-box">
+        <img class="PostIcon" src="{{ asset('storage/user-images/'. $user->images) }}" alt="{{ $user->username }}のアイコン">
+        <div class="postContents">
+             {!! Form::input('text', 'newPost', null, ['required', 'class' => 'form-control', 'placeholder' => '投稿内容を入力してください']) !!}
+        </div>
+        <!-- <button type="submit" class="btn btn-success pull-right"><img class="Upload" src="images/post.png"></button> -->
+        <button type="submit" class="send-btn"><img class="Upload" src="images/post.png"></button>
+
+        <!-- <input type="images" src="{{ asset('images/post.png') }}" class="submit_btn" alt="送信する"> -->
+    </div>
+
  {!! Form::close() !!}
  <table>
         <thead>
             <tr>
-                <th>ユーザー名</th>
-                <th>ユーザーID</th>
-                <th>投稿内容</th>
-                <th>投稿日時</th>
-                <th>編集</th>
-                <th>削除</th>
             </tr>
         </thead>
-     <tbody>
+   <tbody>
 @foreach ($lists as $list)
     <tr>
-        <td>{{ $list->user->username }}</td>
-        <!-- <td>{{ $list->user_id }}</td> -->
-        <td>{{ $list->post }}</td>
-        <td>{{ $list->created_at }}</td>
-        <!-- 更新 -->
-        <td><a class="js-modal-open" href="" post="{{ $list->post }}" post_id="{{ $list->id }}"><img class="Update" src="./images/edit.png" alt="編集" /></a></td>
-        <!-- 削除 -->
-        <td><a class="btn btn-danger" href="/post/{{$list->id}}/delete"onclick="return confirm('この投稿を削除します。よろしいでしょうか？')"><img class="Trash" src="./images/trash.png" alt="削除" /></a></td>
-         <!-- aタグによって別のページに移動する属性でhref属性でその方向性を記述している -->
+        <div class="post-row">
+            <!-- ユーザーのアイコンを表示 -->
+            <td>
+                <img class="MyIcon" src="{{ asset('storage/user-images/'. $list->user->images) }}" alt="{{ $list->user->username }}のアイコン">
+            </td>
+            <td>{{ $list->user->username }}</td>
+            <td>{{ $list->post }}</td>
+            <td>{{ $list->created_at }}</td>
+
+            <!-- 更新 -->
+            @if(Auth::id() == $list->user_id)
+                <td><a class="js-modal-open" href="" post="{{ $list->post }}" post_id="{{ $list->id }}"><img class="Update" src="./images/edit.png" alt="編集" /></a></td>
+            @else
+                <td></td>
+            @endif
+            <!-- 削除 -->
+            @if(Auth::id() == $list->user_id)
+                <td><a class="btn btn-danger" href="/post/{{$list->id}}/delete" onclick="return confirm('この投稿を削除します。よろしいでしょうか？')"><img class="Trash" src="./images/trash.png" alt="削除" /></a></td>
+            @else
+                <td></td>
+            @endif
+
+        </div>
     </tr>
 @endforeach
 </tbody>
